backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 图像处理
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# AI模型
torch==2.1.0
torchvision==0.16.0
clip-by-openai==1.0.0

# 向量数据库
chromadb==0.4.15

# 其他
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# 环境检测
platformdirs==4.1.0